refactor(Clients Page): Moves clients page in the settings menu into the pages folder.
